Fix de l'exception handling et quelque bug

+++ service-publications/src/main/java/com/lms/servicepublications/service/PublicationService.java
@@ -2,64 +2,82 @@ package com.lms.servicepublications.service;
 
 
 import com.lms.servicepublications.dto.PublicationDTO;
-import com.lms.servicepublications.model.Commentaire;
-import com.lms.servicepublications.model.Like;
+import com.lms.servicepublications.exceptions.RessourceNotFoundException;
+import com.lms.servicepublications.exceptions.UnauthorizedException;
 import com.lms.servicepublications.model.Publication;
-import com.lms.servicepublications.repository.CommentaireRepository;
-import com.lms.servicepublications.repository.FichierRepository;
-import com.lms.servicepublications.repository.LikeRepository;
 import com.lms.servicepublications.repository.PublicationRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
-// TODO Vérification de l'identité pour la suppression/Modification
 @Service
 @AllArgsConstructor
 
 public class PublicationService {
 
     private final PublicationRepository publicationRepository;
-    private final FichierRepository fichierRepository;
 
     /**
-     * Fonction pour récupérer une publication
-     * @return
+     * Fonction pour récupérer une publication par son id
+     * @return Publication
      */
     public Publication recupererPublication(String id_publication){
-        Publication publication = publicationRepository.findById(id_publication).orElseThrow(()->new RuntimeException());
-        return publication;
+        return publicationRepository.findById(id_publication).orElseThrow(()->new RessourceNotFoundException("Requested publication is not found."));
     }
 
+    /**
+     * Fonction pour récupérer toutes les publications
+     * @return Publication
+     */
     public List<Publication> recupererPublications(){
-        List<Publication> publications = publicationRepository.findAll();
-        return publications;
+        return publicationRepository.findAll();
     }
+
+    /**
+     * Fonction pour récupérer les publications par cours
+     * @return Publication
+     */
     public List<Publication> recupererPublicationsParCours(String idCours){
-        List<Publication> publications = publicationRepository.findByidCours(idCours);
-        return publications;
+        return publicationRepository.findByidCours(idCours);
     }
-    public void ajouterPublication(PublicationDTO publicationDTO){
+
+    /**
+     * Fonction pour ajouter une publications
+     * @return String
+     */
+    public String ajouterPublication(PublicationDTO publicationDTO){
         Publication publication = new Publication();
         publication.setContenuPublication(publicationDTO.getContenuPublication());
         publication.setIdCours(publicationDTO.getIdCours());
         publication.setIdProprietaire(publicationDTO.getIdProprietaire());
         publicationRepository.insert(publication);
-        System.out.println("publication ajoutée avec succées");
+        return "Publication ajoutée avec succées";
     }
 
-    public void supprimerPublication(String idPublication){
+
+    /**
+     * Fonction pour supprimer une publications par son id
+     * @return String
+     */
+    public String supprimerPublication(String id_user, String idPublication){
+        Publication publication = publicationRepository.findPublicationByid(idPublication);
+        if(!publication.getIdProprietaire().equals(id_user)) throw new UnauthorizedException("Action not authorized");
         publicationRepository.deleteById(idPublication);
-        System.out.println("publication supprimée avec succées");
+        return "Publication supprimée avec succées";
     }
 
-    public Publication modifierPublication(String id,PublicationDTO publicationDTO){
-        Publication publication = publicationRepository.findPublicationByid(id);
+
+    /**
+     * Fonction pour modifier une publications par son id
+     * @return String
+     */
+    public String modifierPublication(String user_id,String publication_id,PublicationDTO publicationDTO){
+        Publication publication = publicationRepository.findById(publication_id).orElseThrow(()->new RessourceNotFoundException("Requested publication is not found."));
+        if(!publication.getIdProprietaire().equals(user_id)) throw new UnauthorizedException("Action not authorized");
         publication.setContenuPublication(publicationDTO.getContenuPublication());
         publication.setIdCours(publicationDTO.getIdCours());
-        publication.setIdProprietaire(publicationDTO.getIdProprietaire());
-        System.out.println("publication modifiée avec succées");
-        return publicationRepository.save(publication);
+        publicationRepository.save(publication);
+        return "Publication modifiée aves succès";
     }
 }
